requirements.txt
flask
pandas
flask-cors
python-dotenv
instaloader
selenium
requests
webdriver-manager
vaderSentiment
browser_cookie3